+++ src/FluentlyHttpClient/Middleware/TimerHttpMiddleware.cs
@@ -6,18 +6,31 @@ using System.Threading.Tasks;
 
 namespace FluentlyHttpClient.Middleware
 {
+	public class TimerHttpMiddlewareOptions
+	{
+		/// <summary>
+		/// Gets or sets the threshold warning timespan in order to log as warning.
+		/// </summary>
+		public TimeSpan WarnThreshold { get; set; } = TimeSpan.FromMilliseconds(250);
+	}
+	
 	/// <summary>
 	/// Timer middleware for Http client.
 	/// </summary>
 	public class TimerHttpMiddleware : IFluentHttpMiddleware
 	{
 		private readonly FluentHttpRequestDelegate _next;
+		private readonly TimerHttpMiddlewareOptions _options;
 		private readonly ILogger _logger;
 
-		public TimerHttpMiddleware(FluentHttpRequestDelegate next, ILogger<TimerHttpMiddleware> logger)
+		public TimerHttpMiddleware(FluentHttpRequestDelegate next, TimerHttpMiddlewareOptions options, ILogger<TimerHttpMiddleware> logger)
 		{
 			_next = next;
+			_options = options;
 			_logger = logger;
+
+			if(_options.WarnThreshold <= TimeSpan.Zero)
+				throw new ArgumentException($"{nameof(_options.WarnThreshold)} must be greater than Zero.");
 		}
 
 		public async Task<FluentHttpResponse> Invoke(FluentHttpRequest request)
@@ -25,10 +38,8 @@ namespace FluentlyHttpClient.Middleware
 			var watch = Stopwatch.StartNew();
 			var response = await _next(request);
 			var elapsed = watch.Elapsed;
-
-			// todo: make configurable
-			const int thresholdMillis = 250;
-			if (_logger.IsEnabled(LogLevel.Warning) && elapsed.TotalMilliseconds >= thresholdMillis)
+			
+			if (_logger.IsEnabled(LogLevel.Warning) && elapsed > _options.WarnThreshold)
 				_logger.LogWarning("Executed request {request} in {timeTakenMillis}ms", request, elapsed.TotalMilliseconds);
 			else if (_logger.IsEnabled(LogLevel.Information))
 				_logger.LogInformation("Executed request {request} in {timeTakenMillis}ms", request, elapsed.TotalMilliseconds);
@@ -60,5 +71,17 @@ namespace FluentlyHttpClient
 		/// <returns>Returns timespan for the time taken.</returns>
 		public static TimeSpan GetTimeTaken(this FluentHttpResponse response)
 			=> (TimeSpan)response.Items[TimeTakenKey];
+	}
+
+	public static class FluentlyHttpMiddlwareExtensions
+	{
+		/// <summary>
+		/// Use timer middleware which measures how long the request takes.
+		/// </summary>
+		/// <param name="builder">Builder instance</param>
+		/// <param name="options">Options to specify for the timer middleware.</param>
+		public static FluentHttpClientBuilder UseTimer(this FluentHttpClientBuilder builder, TimerHttpMiddlewareOptions options = null)
+			=> builder.UseMiddleware<TimerHttpMiddleware>(options ?? new TimerHttpMiddlewareOptions());
+		
 	}
 }