+++ test/TimerHttpMiddlewareTest.cs
@@ -20,8 +20,8 @@ namespace Test
 			var fluentHttpClientFactory = GetNewClientFactory();
 			fluentHttpClientFactory.CreateBuilder("sketch7")
 				.WithBaseUrl("https://sketch7.com")
-				.AddMiddleware<TimerHttpMiddleware>()
 				.WithMessageHandler(mockHttp)
+				.UseTimer()
 				.Register();
 
 			var httpClient = fluentHttpClientFactory.Get("sketch7");
@@ -31,6 +31,22 @@ namespace Test
 			Assert.NotNull(response.Data);
 			Assert.Equal("Azmodan", response.Data.Name);
 			Assert.NotEqual(TimeSpan.Zero, response.GetTimeTaken());
+		}
+
+		[Fact]
+		public async void ThrowsnWhenWarnThresholdIsZero()
+		{
+			var fluentHttpClientFactory = GetNewClientFactory();
+			fluentHttpClientFactory.CreateBuilder("sketch7")
+				.WithBaseUrl("https://sketch7.com")
+				.UseTimer(new TimerHttpMiddlewareOptions
+				{
+					WarnThreshold = TimeSpan.Zero
+				})
+				.Register();
+
+			var httpClient = fluentHttpClientFactory.Get("sketch7");
+			await Assert.ThrowsAsync<ArgumentException>(() => httpClient.Get<Hero>("/api/heroes/azmodan"));
 		}
 	}
 }